+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.9.0] - October 02, 2020
+- Improved readme documentation.
+- Removed material package import.
+## [0.9.0-dev.2] - August 21, 2020
+- Added `isComplete` and `wasComplete` controller getters.
+- Removed `axis` property as it can be handled by `notificationPredicate`.
+## [0.9.0-dev.1] - August 12, 2020
+- Added optional `complete` indicator state together with `completeStateDuration` parameter.
+- `IndicatorControler` changes:
+  - Added `prevoiusState` property.
+  - Added `didStateChange` helper method.
+  - Added `wasArmed`, `wasDragging`, `wasLoading`, `wasHiding` and `wasIdle` properties.
+- Added `notificationPredicate` property to the `CustomRefreshIndicator` widget.
+- Example app:
+  - Added initial version of `check_mark_indicator`. Example that shows how to make use of `complete` state. 
+
 ## [0.8.0+1] - April 12, 2020
 
 ## BREAKING API CHANGES